Introducing replace_edge_anchor -- WIP
