+++ rmf_site_editor/src/interaction/select.rs
@@ -19,21 +19,17 @@ use crate::{
     CurrentWorkspace,
     keyboard::KeyboardServices,
     interaction::*,
-    site::{
-        drawing_editor::CurrentEditDrawing, Anchor, AnchorBundle, Category, CollisionMeshMarker,
-        Dependents, DrawingMarker, Original, PathBehavior, Pending, TextureNeedsAssignment,
-        VisualMeshMarker, ChangeDependent,
-    },
+    site::{drawing_editor::CurrentEditDrawing, Anchor, AnchorBundle, DrawingMarker},
 };
 use rmf_site_format::{
-    Door, Edge, Fiducial, Floor, FrameMarker, Lane, LiftProperties, Location, Measurement, Model,
-    NameInWorkcell, NameOfSite, Path, PixelsPerMeter, Point, Pose, Side, Wall, WorkcellModel,
+    Door, Edge, Lane, LiftProperties, Measurement, NameOfSite, PixelsPerMeter, Wall,
+    Pose,
 };
 use bevy::{
     prelude::{*, Input},
     ecs::system::{SystemParam, StaticSystemParam}
 };
-use bevy_impulse::{*, testing::print_debug};
+use bevy_impulse::*;
 use bevy_mod_raycast::deferred::RaycastMesh;
 use std::{
     collections::HashSet,
@@ -42,7 +38,13 @@ use std::{
 };
 use anyhow::{anyhow, Error as Anyhow};
 
-const SELECT_ANCHOR_MODE_LABEL: &'static str = "select_anchor";
+pub mod create_edges;
+use create_edges::*;
+
+pub mod replace_edge_anchor;
+use replace_edge_anchor::*;
+
+pub const SELECT_ANCHOR_MODE_LABEL: &'static str = "select_anchor";
 
 #[derive(Default)]
 pub struct SelectPlugin {}
@@ -524,7 +526,7 @@ impl AnchorSelectionHelpers {
         update_preview: Service<(Hover, BufferKey<State>), SelectionNodeResult>,
         update_current: Service<(SelectionCandidate, BufferKey<State>), SelectionNodeResult>,
         handle_key_code: Service<(KeyCode, BufferKey<State>), SelectionNodeResult>,
-        cleanup_state: Service<BufferKey<State>, ()>,
+        cleanup_state: Service<BufferKey<State>, SelectionNodeResult>,
         world: &mut World,
     ) -> Service<Option<Entity>, ()> {
         world.spawn_io_workflow(build_anchor_selection_workflow(
@@ -552,23 +554,7 @@ impl AnchorSelectionServices {
         helpers: &AnchorSelectionHelpers,
         app: &mut App,
     ) -> Self {
-        let anchor_setup = app.spawn_service(anchor_selection_setup::<CreateEdges>.into_blocking_service());
-        let state_setup = app.spawn_service(create_edges_setup.into_blocking_service());
-        let update_preview = app.spawn_service(on_hover_anchor_for_edge.into_blocking_service());
-        let update_current = app.spawn_service(on_select_anchor_for_edge.into_blocking_service());
-        let handle_key_code = app.spawn_service(on_esc_for_edge.into_blocking_service());
-        let cleanup_state = app.spawn_service(cleanup_create_edges.into_blocking_service());
-
-        let create_edges = helpers.spawn_anchor_selection_workflow(
-            anchor_setup,
-            state_setup,
-            update_preview,
-            update_current,
-            handle_key_code,
-            cleanup_state,
-            &mut app.world,
-        );
-
+        let create_edges = spawn_create_edges_service(helpers, app);
         Self { create_edges }
     }
 }
@@ -700,7 +686,7 @@ pub fn build_anchor_selection_workflow<State: 'static + Send + Sync>(
     update_preview: Service<(Hover, BufferKey<State>), SelectionNodeResult>,
     update_current: Service<(SelectionCandidate, BufferKey<State>), SelectionNodeResult>,
     handle_key_code: Service<(KeyCode, BufferKey<State>), SelectionNodeResult>,
-    cleanup_state: Service<BufferKey<State>, ()>,
+    cleanup_state: Service<BufferKey<State>, SelectionNodeResult>,
     anchor_cursor_transform: Service<(), ()>,
     anchor_select_stream: Service<(), (), (Hover, Select)>,
     keyboard_just_pressed: Service<(), (), StreamOf<KeyCode>>,
@@ -765,10 +751,17 @@ pub fn build_anchor_selection_workflow<State: 'static + Send + Sync>(
             .connect(scope.terminate);
 
         builder.on_cleanup(buffer, move |scope, builder| {
-            scope.input.chain(builder)
-                .then(cleanup_state)
-                .then(cleanup_anchor_selection)
-                .connect(scope.terminate);
+            let state_node = builder.create_node(cleanup_state);
+            let anchor_node = builder.create_node(cleanup_anchor_selection);
+
+            builder.connect(scope.input, state_node.input);
+            state_node.output.chain(builder)
+                .fork_result(
+                    |ok| ok.connect(anchor_node.input),
+                    |err| err.map_block(print_if_err).connect(anchor_node.input),
+                );
+
+            builder.connect(anchor_node.output, scope.terminate);
         });
     }
 }
@@ -776,393 +769,6 @@ pub fn build_anchor_selection_workflow<State: 'static + Send + Sync>(
 fn print_if_err(err: Option<Anyhow>) {
     if let Some(err) = err {
         error!("{err}");
-    }
-}
-
-pub struct CreateEdges {
-    pub spawn_edge: fn(Edge<Entity>, &mut Commands) -> Entity,
-    pub preview_edge: Option<PreviewEdge>,
-    pub continuity: EdgeContinuity,
-    pub scope: AnchorScope,
-}
-
-impl CreateEdges {
-    pub fn new<T: Bundle + From<Edge<Entity>>>(
-        continuity: EdgeContinuity,
-        scope: AnchorScope,
-    ) -> Self {
-        Self {
-            spawn_edge: create_edge::<T>,
-            preview_edge: None,
-            continuity,
-            scope,
-        }
-    }
-
-    pub fn initialize_preview(
-        &mut self,
-        anchor: Entity,
-        commands: &mut Commands,
-    ) {
-        let edge = Edge::new(anchor, anchor);
-        let edge = (self.spawn_edge)(edge, commands);
-        self.preview_edge = Some(PreviewEdge {
-            edge,
-            side: Side::start(),
-            provisional_start: false,
-        });
-
-        commands.add(ChangeDependent::add(anchor, edge));
-    }
-}
-
-fn create_edge<T: Bundle + From<Edge<Entity>>>(
-    edge: Edge<Entity>,
-    commands: &mut Commands,
-) -> Entity {
-    let new_bundle: T = edge.into();
-    commands.spawn((new_bundle, Pending)).id()
-}
-
-#[derive(Clone, Copy)]
-pub struct PreviewEdge {
-    pub edge: Entity,
-    pub side: Side,
-    /// True if the start anchor of the edge was created specifically to build
-    /// this edge. If this true, we will despawn the anchor during cleanup if
-    /// the edge does not get completed.
-    pub provisional_start: bool,
-}
-
-impl PreviewEdge {
-    pub fn cleanup(
-        &self,
-        edges: &Query<&'static Edge<Entity>>,
-        commands: &mut Commands,
-    ) -> SelectionNodeResult {
-        let edge = edges.get(self.edge).or_broken_query()?;
-        for anchor in edge.array() {
-            commands.add(ChangeDependent::remove(anchor, self.edge));
-        }
-
-        if self.provisional_start {
-            // The start anchor was created specifically for this preview edge
-            // which we are about to despawn. Let's despawn both so we aren't
-            // littering the scene with unintended anchors.
-            commands.get_entity(edge.start()).or_broken_query()?.despawn_recursive();
-        }
-
-        commands.get_entity(self.edge).or_broken_query()?.despawn_recursive();
-        Ok(())
-    }
-}
-
-pub enum EdgeContinuity {
-    /// Create just a single edge
-    Single,
-    /// Create a sequence of separate edges
-    Separate,
-    /// Create edges continuously, i.e. the beginning of the next edge will
-    /// automatically be the end of the previous edge.
-    Continuous,
-}
-
-impl Borrow<AnchorScope> for CreateEdges {
-    fn borrow(&self) -> &AnchorScope {
-        &self.scope
-    }
-}
-
-pub fn anchor_selection_setup<State: Borrow<AnchorScope>>(
-    In(key): In<BufferKey<State>>,
-    access: BufferAccess<State>,
-    anchors: Query<Entity, With<Anchor>>,
-    drawings: Query<(), With<DrawingMarker>>,
-    parents: Query<&'static Parent>,
-    mut visibility: Query<&'static mut Visibility>,
-    mut hidden_anchors: ResMut<HiddenSelectAnchorEntities>,
-    mut current_anchor_scope: ResMut<AnchorScope>,
-    mut cursor: ResMut<Cursor>,
-    mut highlight: ResMut<HighlightAnchors>,
-) -> SelectionNodeResult
-where
-    State: 'static + Send + Sync,
-{
-    let access = access.get(&key).or_broken_buffer()?;
-    let state = access.newest().or_missing_state()?;
-    let scope: &AnchorScope = (&*state).borrow();
-    match scope {
-        AnchorScope::General | AnchorScope::Site => {
-            // If we are working with normal level or site requests, hide all drawing anchors
-            for e in anchors.iter().filter(|e| {
-                parents
-                .get(*e)
-                .is_ok_and(|p| drawings.get(p.get()).is_ok())
-            }) {
-                set_visibility(e, &mut visibility, false);
-                hidden_anchors.drawing_anchors.insert(e);
-            }
-        }
-        // Nothing to hide, it's done by the drawing editor plugin
-        AnchorScope::Drawing => {}
-    }
-
-    if scope.is_site() {
-        set_visibility(cursor.site_anchor_placement, &mut visibility, true);
-    } else {
-        set_visibility(cursor.level_anchor_placement, &mut visibility, true);
-    }
-
-    highlight.0 = true;
-
-    *current_anchor_scope = *scope;
-
-    cursor.add_mode(SELECT_ANCHOR_MODE_LABEL, &mut visibility);
-
-    Ok(())
-}
-
-pub fn create_edges_setup(
-    In(key): In<BufferKey<CreateEdges>>,
-    mut access: BufferAccessMut<CreateEdges>,
-    cursor: Res<Cursor>,
-    mut commands: Commands,
-) -> SelectionNodeResult {
-    let mut access = access.get_mut(&key).or_broken_buffer()?;
-    let state = access.newest_mut().or_missing_state()?;
-
-    if state.preview_edge.is_none() {
-        state.initialize_preview(cursor.level_anchor_placement, &mut commands);
-    }
-    Ok(())
-}
-
-pub fn on_hover_anchor_for_edge(
-    In((hover, key)): In<(Hover, BufferKey<CreateEdges>)>,
-    mut access: BufferAccessMut<CreateEdges>,
-    mut cursor: ResMut<Cursor>,
-    mut visibility: Query<&mut Visibility>,
-    mut edges: Query<&mut Edge<Entity>>,
-    mut commands: Commands,
-) -> SelectionNodeResult {
-    let mut access = access.get_mut(&key).or_broken_buffer()?;
-    let state = access.newest_mut().or_missing_state()?;
-
-    let anchor = match hover.0 {
-        Some(anchor) => {
-            cursor.remove_mode(SELECT_ANCHOR_MODE_LABEL, &mut visibility);
-            anchor
-        }
-        None => {
-            cursor.add_mode(SELECT_ANCHOR_MODE_LABEL, &mut visibility);
-            cursor.level_anchor_placement
-        }
-    };
-
-    if let Some(preview) = &mut state.preview_edge {
-        // If we already have an active preview, then use the new anchor for the
-        // side that we currently need to select for.
-        let index = preview.side.index();
-        let mut edge = edges.get_mut(preview.edge).or_broken_query()?;
-
-        let old_anchor = edge.array()[index];
-        if old_anchor != anchor {
-            let opposite_anchor = edge.array()[preview.side.opposite().index()];
-            if opposite_anchor != old_anchor {
-                commands.add(ChangeDependent::remove(old_anchor, preview.edge));
-            }
-
-            edge.array_mut()[index] = anchor;
-            commands.add(ChangeDependent::add(anchor, preview.edge));
-        }
-    } else {
-        // There is currently no active preview, so we need to create one.
-        let edge = Edge::new(anchor, anchor);
-        let edge = (state.spawn_edge)(edge, &mut commands);
-        state.preview_edge = Some(PreviewEdge {
-            edge,
-            side: Side::start(),
-            provisional_start: false,
-        });
-        commands.add(ChangeDependent::add(anchor, edge));
-    }
-
-    Ok(())
-}
-
-pub fn on_select_anchor_for_edge(
-    In((selection, key)): In<(SelectionCandidate, BufferKey<CreateEdges>)>,
-    mut access: BufferAccessMut<CreateEdges>,
-    mut edges: Query<&mut Edge<Entity>>,
-    mut commands: Commands,
-    cursor: Res<Cursor>,
-) -> SelectionNodeResult {
-    let mut access = access.get_mut(&key).or_broken_buffer()?;
-    let state = access.newest_mut().or_missing_state()?;
-
-    let anchor = selection.candidate;
-    if let Some(preview) = &mut state.preview_edge {
-        match preview.side {
-            Side::Left => {
-                // We are pinning down the first anchor of the edge
-                let mut edge = edges.get_mut(preview.edge).or_broken_query()?;
-                commands.add(ChangeDependent::remove(edge.left(), preview.edge));
-                *edge.left_mut() = anchor;
-                commands.add(ChangeDependent::add(anchor, preview.edge));
-
-                if edge.right() != anchor {
-                    commands.add(ChangeDependent::remove(edge.right(), preview.edge));
-                }
-
-                *edge.right_mut() = cursor.level_anchor_placement;
-                commands.add(ChangeDependent::add(cursor.level_anchor_placement, preview.edge));
-
-                preview.side = Side::Right;
-                preview.provisional_start = selection.provisional;
-            }
-            Side::Right => {
-                // We are finishing the edge
-                let mut edge = edges.get_mut(preview.edge).or_broken_query()?;
-                if edge.left() == anchor {
-                    // The user is trying to use the same point for the start
-                    // and end of an edge. Issue a warning and exit early.
-                    warn!(
-                        "You are trying to select an anchor {:?} for both the \
-                        start and end points of an edge, which is not allowed.",
-                        anchor,
-                    );
-                    return Ok(());
-                }
-                *edge.right_mut() = anchor;
-                commands.add(ChangeDependent::add(anchor, preview.edge));
-                commands.get_entity(preview.edge).or_broken_query()?.remove::<Preview>();
-
-                match state.continuity {
-                    EdgeContinuity::Single => {
-                        state.preview_edge = None;
-                        // This simply means we are terminating the workflow now
-                        // because we have finished drawing the single edge
-                        return Err(None);
-                    }
-                    EdgeContinuity::Separate => {
-                        // Start drawing a new edge from a blank slate with the
-                        // next selection
-                        state.initialize_preview(cursor.level_anchor_placement, &mut commands);
-                    }
-                    EdgeContinuity::Continuous => {
-                        // Start drawing a new edge, picking up from the end
-                        // point of the previous edge
-                        let edge = Edge::new(anchor, cursor.level_anchor_placement);
-                        let edge = (state.spawn_edge)(edge, &mut commands);
-                        state.preview_edge = Some(PreviewEdge {
-                            edge,
-                            side: Side::end(),
-                            provisional_start: false,
-                        });
-                        commands.add(ChangeDependent::add(anchor, edge));
-                        commands.add(ChangeDependent::add(cursor.level_anchor_placement, edge));
-                    }
-                }
-            }
-        }
-    } else {
-        // We have no preview at all yet somehow, so we'll need to create a
-        // fresh new edge to insert the selected anchor into
-        let edge = Edge::new(anchor, anchor);
-        let edge = (state.spawn_edge)(edge, &mut commands);
-        state.preview_edge = Some(PreviewEdge {
-            edge,
-            side: Side::start(),
-            provisional_start: selection.provisional,
-        });
-    }
-
-    Ok(())
-}
-
-pub fn on_esc_for_edge(
-    In((button, key)): In<(KeyCode, BufferKey<CreateEdges>)>,
-    mut access: BufferAccessMut<CreateEdges>,
-    mut edges: Query<&'static mut Edge<Entity>>,
-    cursor: Res<Cursor>,
-    mut commands: Commands,
-) -> SelectionNodeResult {
-    if !matches!(button, KeyCode::Escape) {
-        // The button was not the escape key, so there's nothing for us to do
-        // here.
-        return Ok(());
-    }
-
-    let mut access = access.get_mut(&key).or_broken_buffer()?;
-    let state = access.newest_mut().or_missing_state()?;
-
-    if let Some(preview) = &mut state.preview_edge {
-        if preview.side == Side::end() {
-            // We currently have an active preview edge and are selecting for
-            // the second point in the edge. Esc means we should back out of the
-            // current edge without exiting the edge creation workflow so the
-            // user can choose a different start point.
-            let mut edge = edges.get_mut(preview.edge).or_broken_query()?;
-            for anchor in edge.array() {
-                commands.add(ChangeDependent::remove(anchor, preview.edge));
-            }
-            if preview.provisional_start {
-                commands.get_entity(edge.start()).or_broken_query()?.despawn_recursive();
-            }
-
-            *edge.left_mut() = cursor.level_anchor_placement;
-            *edge.right_mut() = cursor.level_anchor_placement;
-            preview.side = Side::start();
-            preview.provisional_start = false;
-            commands.add(ChangeDependent::add(cursor.level_anchor_placement, preview.edge));
-
-        } else {
-            // We are selecting for the first point in the edge. If the user has
-            // pressed Esc then that means they want to stop creating edges
-            // altogether. Return Err(None) to indicate that the workflow should
-            // exit cleaning.
-            return Err(None);
-        }
-    } else {
-        // We currently have no preview active at all. If the user hits Esc then
-        // they want to exit the workflow altogether.
-        return Err(None);
-    }
-
-    Ok(())
-}
-
-pub fn cleanup_create_edges(
-    In(key): In<BufferKey<CreateEdges>>,
-    mut access: BufferAccessMut<CreateEdges>,
-    edges: Query<&'static Edge<Entity>>,
-    mut commands: Commands,
-) {
-    let mut access = match access.get_mut(&key).or_broken_buffer() {
-        Ok(access) => access,
-        Err(err) => {
-            if let Some(err) = err {
-                error!("{err}");
-            }
-            return;
-        }
-    };
-
-    let state = match access.pull().or_missing_state() {
-        Ok(state) => state,
-        Err(err) => {
-            if let Some(err) = err {
-                error!("{err}");
-            }
-            return;
-        }
-    };
-
-    if let Some(preview) = state.preview_edge {
-        // We created a preview, so we should despawn it while cleaning up
-        if let Err(Some(err)) = preview.cleanup(&edges, &mut commands) {
-            error!("{err}");
-        }
     }
 }
 
@@ -1286,7 +892,7 @@ impl<'w, 's> SelectionFilter for AnchorFilter<'w ,'s> {
                     error!("Cannot find pixels per meter of current drawing");
                     return None;
                 };
-                let pose = compute_parent_inverse_pose(&tf, &self.transforms, drawing);
+                let pose = compute_parent_inverse_pose(&tf, &self.transforms, drawing)?;
                 let ppm = ppm.0;
                 self.commands
                     .spawn(AnchorBundle::new([pose.trans[0], pose.trans[1]].into()))
@@ -1309,16 +915,20 @@ fn compute_parent_inverse_pose(
     tf: &GlobalTransform,
     transforms: &Query<&GlobalTransform>,
     parent: Entity,
-) -> Pose {
-    let parent_tf = transforms
-        .get(parent)
-        .expect("Failed in fetching parent transform");
+) -> Option<Pose> {
+    let parent_tf = match transforms.get(parent) {
+        Ok(tf) => tf,
+        Err(err) => {
+            error!("Failed in fetching parent transform: {err}");
+            return None;
+        }
+    };
 
     let inv_tf = parent_tf.affine().inverse();
     let goal_tf = tf.affine();
     let mut pose = Pose::default();
     pose.rot = pose.rot.as_euler_extrinsic_xyz();
-    pose.align_with(&Transform::from_matrix((inv_tf * goal_tf).into()))
+    Some(pose.align_with(&Transform::from_matrix((inv_tf * goal_tf).into())))
 }
 
 pub trait SelectionFilter: SystemParam {
@@ -1574,124 +1184,51 @@ pub fn clear_hover_select(
     }
 }
 
-// pub fn handle_selection_picking(
-//     blockers: Option<Res<SelectionBlockers>>,
-//     mode: Res<InteractionMode>,
-//     selectables: Query<&Selectable>,
-//     anchors: Query<(), With<Anchor>>,
-//     mut picks: EventReader<ChangePick>,
-//     mut hover: EventWriter<Hover>,
-// ) {
-//     if let Some(blockers) = blockers {
-//         if blockers.blocking() {
-//             hover.send(Hover(None));
-//             return;
-//         }
-//     }
+pub fn anchor_selection_setup<State: Borrow<AnchorScope>>(
+    In(key): In<BufferKey<State>>,
+    access: BufferAccess<State>,
+    anchors: Query<Entity, With<Anchor>>,
+    drawings: Query<(), With<DrawingMarker>>,
+    parents: Query<&'static Parent>,
+    mut visibility: Query<&'static mut Visibility>,
+    mut hidden_anchors: ResMut<HiddenSelectAnchorEntities>,
+    mut current_anchor_scope: ResMut<AnchorScope>,
+    mut cursor: ResMut<Cursor>,
+    mut highlight: ResMut<HighlightAnchors>,
+) -> SelectionNodeResult
+where
+    State: 'static + Send + Sync,
+{
+    let access = access.get(&key).or_broken_buffer()?;
+    let state = access.newest().or_missing_state()?;
+    let scope: &AnchorScope = (&*state).borrow();
+    match scope {
+        AnchorScope::General | AnchorScope::Site => {
+            // If we are working with normal level or site requests, hide all drawing anchors
+            for e in anchors.iter().filter(|e| {
+                parents
+                .get(*e)
+                .is_ok_and(|p| drawings.get(p.get()).is_ok())
+            }) {
+                set_visibility(e, &mut visibility, false);
+                hidden_anchors.drawing_anchors.insert(e);
+            }
+        }
+        // Nothing to hide, it's done by the drawing editor plugin
+        AnchorScope::Drawing => {}
+    }
 
-//     if !mode.is_selecting() {
-//         hover.send(Hover(None));
-//         return;
-//     }
+    if scope.is_site() {
+        set_visibility(cursor.site_anchor_placement, &mut visibility, true);
+    } else {
+        set_visibility(cursor.level_anchor_placement, &mut visibility, true);
+    }
 
-//     for pick in picks.read() {
-//         hover.send(Hover(
-//             pick.to
-//                 .and_then(|change_pick_to| {
-//                     selectables
-//                         .get(change_pick_to)
-//                         .ok()
-//                         .map(|selectable| selectable.element)
-//                 })
-//                 .and_then(|change_pick_to| {
-//                     if let InteractionMode::SelectAnchor(_) = *mode {
-//                         if anchors.contains(change_pick_to) {
-//                             Some(change_pick_to)
-//                         } else {
-//                             None
-//                         }
-//                     } else {
-//                         Some(change_pick_to)
-//                     }
-//                 }),
-//         ));
-//     }
-// }
+    highlight.0 = true;
 
-// pub fn maintain_hovered_entities(
-//     mut hovered: Query<&mut Hovered>,
-//     mut hovering: ResMut<Hovering>,
-//     mut hover: EventReader<Hover>,
-//     mouse_button_input: Res<Input<MouseButton>>,
-//     touch_input: Res<Touches>,
-//     mut select: EventWriter<Select>,
-//     mode: Res<InteractionMode>,
-//     blockers: Option<Res<PickingBlockers>>,
-// ) {
-//     if let Some(new_hovered) = hover.read().last() {
-//         if hovering.0 != new_hovered.0 {
-//             if let Some(previous_hovered) = hovering.0 {
-//                 if let Ok(mut hovering) = hovered.get_mut(previous_hovered) {
-//                     hovering.is_hovered = false;
-//                 }
-//             }
+    *current_anchor_scope = *scope;
 
-//             if let Some(new_hovered) = new_hovered.0 {
-//                 if let Ok(mut hovering) = hovered.get_mut(new_hovered) {
-//                     hovering.is_hovered = true;
-//                 }
-//             }
+    cursor.add_mode(SELECT_ANCHOR_MODE_LABEL, &mut visibility);
 
-//             hovering.0 = new_hovered.0;
-//         }
-//     }
-
-//     let clicked = mouse_button_input.just_pressed(MouseButton::Left)
-//         || touch_input.iter_just_pressed().next().is_some();
-//     let blocked = blockers.filter(|x| x.blocking()).is_some();
-
-//     if clicked && !blocked {
-//         if let Some(current_hovered) = hovering.0 {
-//             // TODO(luca) refactor to remove this hack
-//             // Skip if we are in SelectAnchor3D mode
-//             if let InteractionMode::SelectAnchor3D(_) = &*mode {
-//                 return;
-//             }
-//             select.send(Select(Some(current_hovered)));
-//         }
-//     }
-// }
-
-// pub fn maintain_selected_entities(
-//     mode: Res<InteractionMode>,
-//     mut selected: Query<&mut Selected>,
-//     mut selection: ResMut<Selection>,
-//     mut select: EventReader<Select>,
-// ) {
-//     if !mode.is_inspecting() {
-//         // We only maintain the "selected" entity when we are in Inspect mode.
-//         // Other "selecting" modes, like SelectAnchor, take in the selection as
-//         // an event and do not change the current selection that is being
-//         // inspected.
-//         return;
-//     }
-
-//     if let Some(new_selection) = select.read().last() {
-//         if selection.0 != new_selection.0 {
-//             if let Some(previous_selection) = selection.0 {
-//                 if let Ok(mut selected) = selected.get_mut(previous_selection) {
-//                     selected.is_selected = false;
-//                 }
-//             }
-
-//             if let Some(new_selection) = new_selection.0 {
-//                 if let Ok(mut selected) = selected.get_mut(new_selection) {
-//                     selected.is_selected = true;
-//                 }
-//             }
-
-//             selection.0 = new_selection.0;
-//         }
-//     }
-// }
-
+    Ok(())
+}
